arch: Move up_getsp from arch.h to irq.h
since all other special register operation in irq.h
