+++ arch/arm/include/arch.h
@@ -40,8 +40,6 @@
  * Pre-processor Prototypes
  ****************************************************************************/
 
-#define up_getsp()              (uintptr_t)__builtin_frame_address(0)
-
 #ifdef CONFIG_PIC
 
 /* This identifies the register the is used by the processor as the PIC base
