+++ arch/arm/include/irq.h
@@ -60,6 +60,12 @@
 #  include <arch/arm/irq.h>
 #endif
 
+/****************************************************************************
+ * Pre-processor Prototypes
+ ****************************************************************************/
+
+#define up_getsp() (uintptr_t)__builtin_frame_address(0)
+
 #ifndef __ASSEMBLY__
 
 #ifdef __cplusplus
