+++ arch/avr/include/avr/arch.h
@@ -40,30 +40,6 @@
 #ifndef __ASSEMBLY__
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint16_t up_getsp(void)
-{
-  uint8_t spl;
-  uint8_t sph;
-
-  __asm__ __volatile__
-  (
-    "in %0, __SP_L__\n\t"
-    "in %1, __SP_H__\n"
-    : "=r" (spl), "=r" (sph)
-    :
-  );
-
-  return (uint16_t)sph << 8 | spl;
-}
-
-/****************************************************************************
  * Public Variables
  ****************************************************************************/
