+++ arch/avr/include/avr32/arch.h
@@ -40,27 +40,6 @@
 #ifndef __ASSEMBLY__
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint32_t up_getsp(void)
-{
-  uint32_t retval;
-  __asm__ __volatile__
-    (
-      "mov\t%0,sp\n\t"
-      : "=r" (retval)
-      :
-    );
-
-  return retval;
-}
-
-/****************************************************************************
  * Public Variables
  ****************************************************************************/
