+++ arch/hc/include/arch.h
@@ -36,25 +36,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint16_t up_getsp(void)
-{
-  uint16_t ret;
-  __asm__
-  (
-    "\tsts %0\n"
-  : "=m"(ret) :
-  );
-  return ret;
-}
-
-/****************************************************************************
  * Public Types
  ****************************************************************************/
