+++ arch/hc/include/irq.h
@@ -70,6 +70,23 @@ extern "C"
 #endif
 
 /****************************************************************************
+ * Inline functions
+ ****************************************************************************/
+
+/* Return the current value of the stack pointer */
+
+static inline uint16_t up_getsp(void)
+{
+  uint16_t ret;
+  __asm__
+  (
+    "\tsts %0\n"
+  : "=m"(ret) :
+  );
+  return ret;
+}
+
+/****************************************************************************
  * Public Data
  ****************************************************************************/
