+++ arch/misoc/include/arch.h
@@ -29,23 +29,8 @@
  * Included Files
  ****************************************************************************/
 
-#include <nuttx/config.h>
-
 /****************************************************************************
  * Inline functions
  ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint32_t up_getsp(void)
-{
-  register uint32_t sp;
-
-  __asm__ __volatile__("addi %0, sp, 0" : "=r" (sp));
-
-  return sp;
-}
 
 #endif /* __ARCH_MISOC_INCLUDE_ARCH_H */
