+++ arch/misoc/include/irq.h
@@ -58,6 +58,21 @@ extern "C"
 #endif
 
 /****************************************************************************
+ * Inline functions
+ ****************************************************************************/
+
+/* Return the current value of the stack pointer */
+
+static inline uint32_t up_getsp(void)
+{
+  register uint32_t sp;
+
+  __asm__ __volatile__("addi %0, sp, 0" : "=r" (sp));
+
+  return sp;
+}
+
+/****************************************************************************
  * Public Data
  ****************************************************************************/
