+++ arch/renesas/include/rx65n/arch.h
@@ -40,29 +40,6 @@
 #ifndef __ASSEMBLY__
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint16_t up_getsp(void)
-{
-  uint16_t sp;
-
-  __asm__ __volatile__
-    (
-      "\tmvfc usp, %0\n\t"
-      : "=r" (sp)
-      :
-      :"memory"
-    );
-
-  return sp;
-}
-
-/****************************************************************************
  * Public Variables
  ****************************************************************************/
