+++ arch/risc-v/include/arch.h
@@ -79,29 +79,6 @@
 #endif
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-#ifndef __ASSEMBLY__
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uintptr_t up_getsp(void)
-{
-  register uintptr_t sp;
-  __asm__
-  (
-    "\tadd  %0, x0, x2\n"
-    : "=r"(sp)
-  );
-  return sp;
-}
-
-#endif
-
-/****************************************************************************
  * Public Types
  ****************************************************************************/
