+++ arch/risc-v/include/irq.h
@@ -569,6 +569,23 @@ struct xcptcontext
 #ifndef __ASSEMBLY__
 
 /****************************************************************************
+ * Inline functions
+ ****************************************************************************/
+
+/* Return the current value of the stack pointer */
+
+static inline uintptr_t up_getsp(void)
+{
+  register uintptr_t sp;
+  __asm__
+  (
+    "\tadd  %0, x0, x2\n"
+    : "=r"(sp)
+  );
+  return sp;
+}
+
+/****************************************************************************
  * Public Data
  ****************************************************************************/
