+++ arch/sparc/include/arch.h
@@ -36,28 +36,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint32_t up_getsp(void)
-{
-  uint32_t retval;
-
-  do
-    {
-      retval = 0;
-      __asm__ volatile("mov %%sp, %0" :  "=r" (retval) : "0" (retval));
-    }
-  while (0);
-
-  return retval;
-}
-
-/****************************************************************************
  * Public Types
  ****************************************************************************/
