+++ arch/sparc/include/irq.h
@@ -69,6 +69,26 @@ extern "C"
 #endif
 
 /****************************************************************************
+ * Inline functions
+ ****************************************************************************/
+
+/* Return the current value of the stack pointer */
+
+static inline uint32_t up_getsp(void)
+{
+  uint32_t retval;
+
+  do
+    {
+      retval = 0;
+      __asm__ volatile("mov %%sp, %0" :  "=r" (retval) : "0" (retval));
+    }
+  while (0);
+
+  return retval;
+}
+
+/****************************************************************************
  * Public Data
  ****************************************************************************/
