+++ arch/xtensa/include/arch.h
@@ -40,28 +40,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Inline functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: up_getsp
- ****************************************************************************/
-
-static inline uint32_t up_getsp(void) inline_function;
-static inline uint32_t up_getsp(void)
-{
-  register uint32_t sp;
-
-  __asm__ __volatile__
-  (
-    "mov %0, sp\n"
-    : "=r" (sp)
-  );
-
-  return sp;
-}
-
-/****************************************************************************
  * Public Types
  ****************************************************************************/
