+++ arch/z16/include/arch.h
@@ -59,10 +59,6 @@ extern "C"
 #define EXTERN extern
 #endif
 
-/* Return the current value of the stack pointer */
-
-chipreg_t up_getsp(void);
-
 #undef EXTERN
 #ifdef __cplusplus
 }
