+++ arch/z16/include/irq.h
@@ -62,6 +62,10 @@ extern "C"
 #define EXTERN extern
 #endif
 
+/* Return the current value of the stack pointer */
+
+chipreg_t up_getsp(void);
+
 /****************************************************************************
  * Public Data
  ****************************************************************************/
