+++ arch/z80/include/irq.h
@@ -32,4 +32,25 @@
 #include <nuttx/irq.h>
 #include <arch/chip/irq.h>
 
+/****************************************************************************
+ * Public Function Prototypes
+ ****************************************************************************/
+
+#ifdef __cplusplus
+#define EXTERN extern "C"
+extern "C"
+{
+#else
+#define EXTERN extern
+#endif
+
+/* Return the current value of the stack pointer */
+
+uintptr_t up_getsp(void);
+
+#undef EXTERN
+#ifdef __cplusplus
+}
+#endif
+
 #endif /* __ARCH_Z80_INCLUDE_IRQ_H */
